fixed categoryService

+++ src/main/java/com/example/moneymissint/service/CategoryService.java
@@ -1,5 +1,4 @@
 package com.example.moneymissint.service;
-
 import com.example.moneymissint.model.Category;
 import com.example.moneymissint.model.User;
 import com.example.moneymissint.repository.CategoryRepository;
@@ -11,18 +10,16 @@ import jakarta.validation.constraints.NotBlank;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.validation.annotation.Validated;
 
 @Service
 @RequiredArgsConstructor
 @Transactional
+@Validated
 public class CategoryService {
 
     private final CategoryRepository categoryRepository;
-
-    private final UserRepository userRepository;
-
     private final TransactionRepository transactionRepository;
-
     private final UserService userService;
 
     public Category getCategoryOrThrow(Long categoryId) {
@@ -35,22 +32,41 @@ public class CategoryService {
         if (user.getStatus() != Status.ACTIVE){
             throw new IllegalStateException("User is inactive");
         }
-
+        if (categoryRepository.existsByUserIdAndCategoryNameIgnoreCase(userId, categoryName.trim())){
+            throw new IllegalStateException("Category already exists");
+        }
         Category category = new Category();
         category.setCategoryName(categoryName);
         category.setUser(user);
         return categoryRepository.save(category);
     }
 
-    public Category updateCategory(Long categoryId, String categoryName){
+    public Category renameCategory(Long categoryId, @NotBlank String categoryName, Long userId){
+
         Category category = getCategoryOrThrow(categoryId);
-        category.setCategoryName(categoryName);
+        if (!category.getUser().getId().equals(userId)) {
+            throw new IllegalStateException("Category does not belong to user");
+        }
+
+        String normalizedCategoryName = categoryName.trim();
+
+        if (categoryRepository.existsByUserIdAndCategoryNameIgnoreCaseAndIdNot(userId, normalizedCategoryName, categoryId)){
+            throw new IllegalStateException("Category already exists");
+        }
+
+
+        category.setCategoryName(normalizedCategoryName);
 
         return category;
     }
 
-    public void  deleteCategory(Long categoryId){
+    public void  deleteCategory(Long userId, Long categoryId){
         Category category = getCategoryOrThrow(categoryId);
+        if (!category.getUser().getId().equals(userId)){
+            throw new IllegalStateException("Category does not belong to user");
+        }
+
+        categoryRepository.clearCategoryByCategoryId(categoryId);
         categoryRepository.delete(category);
     }
